Refactor req->new-activity into new-activity
This simplifies the activity namespace and leaves all the logic to the handlers
namespace, which acts as the edge of the application.

Centralizing the effects of calling specific handlers in the handlers namespace
itself improves legibility and decreases the risk of passing down improperly
transformed data deeper into our application.