Refactor path to simplify access to routes in different namespaces
The previous version took the entire request and extracted the router from it.
Now path takes only the router, the route name and optional extra params. This
way the logic (that the router is included in every request) behind reitit path
building becomes more visible and the data flow is simplified as we don't have
to pass the entire request to every function that needs access to the router.